+++ recipes/configure-ua-harvester.rb
@@ -13,14 +13,6 @@ sqs_queue_name = "#{stack_name}-user-actions"
 region = node[:opsworks][:instance][:region]
 es_host = node[:opsworks][:instance][:private_ip]
 max_start_end_span = elk_info.fetch('harvester_max_start_end_span', nil)
-zoom_key = elk_info['zoom_key']
-zoom_secret = elk_info['zoom_secret']
-loggly_info = node.fetch(:loggly, { token: '', url: '' })
-loggly_config = if loggly_info[:token] != ''
-                  %Q|LOGGLY_TOKEN=#{loggly_info[:token]}|
-                else
-                  ''
-                end
 
 file '/home/ua_harvester/harvester/.env' do
   owner 'ua_harvester'
@@ -32,17 +24,11 @@ MATTERHORN_REST_PASS="#{rest_auth_info[:pass]}"
 MATTERHORN_ENGAGE_HOST="#{engage_node[:private_ip]}"
 MATTERHORN_ADMIN_HOST="#{admin_node[:private_ip]}"
 ES_HOST="#{es_host}"
-ZOOM_KEY=#{zoom_key}
-ZOOM_SECRET=#{zoom_secret}
 GEOLITE_PATH=/opt/geolite2/GeoLite2-City.mmdb
 S3_HARVEST_TS_BUCKET="#{stack_name}-ua-harvester"
 S3_LAST_ACTION_TS_KEY="#{stack_name}-last-action-ts"
 SQS_QUEUE_NAME="#{sqs_queue_name}"
 MAX_START_END_SPAN=#{max_start_end_span}
-LOGGLY_TAGS="#{stack_name}"
-#{loggly_config}
 |
   mode '600'
 end
-
-